CART-89 hash: Fix d_chash_rec_first()
When porting this code from IOF I changed a < 0 to a != 0 but didn't test
it.  The hash table code returns a negative number on error but the return
value of the callback on non-zero.  In this case we're using the callback
to abort the traverse so the callback returns 1, and with the change to use
!= 0 the first function was incorrectly using this to return NULL rather
than rlink.

The net effect of this is that first() was always returning NULL, even
when there were entries in the hash table.

Change-Id: I60a14e8d21346d5879d231281524258484349df4
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29298
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>