CORCI-126 build: Use installed components if desired
--build-deps=yes causes scons_local to always build a
prerequisite if it can be built.   This is useful behavior when
we definitely want to ignore a system library.  However, sometimes,
maybe more often than not at this point, we want to use the
installed one instead.  Add an option that allows us to change this
behavior for any or all prerequisites.

Change-Id: Ifb3d7f03d922e3ca51460f155aebdd13b6f8595d
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/29357
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>