CORCI-127 Fix update_devel_branch.sh
The cart_devel branch is for the Jenkins matrix jobs to pull the commit
hash from the cart build.config.

The same is for iof_devel, cppr_devel, and cppr_devel branches.

The force push is needed to update the designated branch if Gerrit
decides that the new branch is not a direct ancestor of the previous
branch commit.

If the update_branch Jenkins jobs for cart/iof/cppr/daos fail, then the
*-devel Jenkins jobs will not be building the requested commits for
those jobs.

Change-Id: I36930246778aac3b5184bd925971575933127f1d
Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>
Reviewed-on: https://review.whamcloud.com/29533
Tested-by: Jenkins
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>