CART-334 addref: Overhaul addref/decref handling to be consistent.
Add RPC_(ADD|DEC)REF macros to modify the reference count directly
if a rpc_priv pointer is held, this is slightly more efficient but
dramatially improves logging.

Error checking was rarely applied to the functions, and where it was
typically it resulted in an additional log message only (in addition
to the one from the called function) so simply remove all these checks
as rpc_priv should already have been checked (null RPC was the only
failure the functions checked for anyway).

Change-Id: I7da8071db62cf706c39d78b39c13a03731498ecf
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29819
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>