CART-337 corpc: add CRT_RPC_FLAG_GRP_DESTROY flag handling
Can set a CRT_RPC_FLAG_GRP_DESTROY for corpc request in that case will
destroy the group together with that bcast's completion.
This can save one bcast in some cases, for example in daos pool destroy
now it firstly bcast a RDP_STOP rpc and then destroy the pool group, it
can be refine by piggyback a CRT_RPC_FLAG_GRP_DESTROY flag for the first
bcast.

And refined the corpc completion and error handling.

Change-Id: Id4a17514a5626e81cd13989f9e183f37dc5ab1e2
Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>
Reviewed-on: https://review.whamcloud.com/29842
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>