DAOS-387 server: Add setup/cleanup module callbacks
This patch adds a new pair of server module callbacks,
sm_setup/sm_cleanup, to support functionalities like starting existing
pool services when starting servers and properly shutting down modules
when stopping servers.

An implementation of sm_cleanup for pool is provided to stop all pool
services, so that shutting down a server does not hang for the rdb
daemons.

server_fini()'s finalization order is also fixed to be in the reverse
order of server_init().

Change-Id: Ib8c44741d4b05e254368d5eb7b108a5dbf9d1b7e
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.whamcloud.com/29876
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>