DAOS-372 test: initial patch for pool functional tests.
This patch includes some initial work to create a set of functional
tests for DAOS_M.  This initial check-in includes a lot of temporary and
speculative code that will be removed/altered in subsequent patches.  
This test code uses the Avocado test framework.  This must be yum 
installed on all test systems prior to attempting to run these tests.  

All the tests can be individually run using avocado but this requires
knowledge of the tests and avocado options.  The patch includes a script
runtest.bash which can be used to launch the tests.  

"runtest.bash -p" can be used to trigger all the pool related tests 
included in this patch.

Change-Id: I1a4c9d120e826e54e0be4716b9c694559721a9bb
Signed-off-by: Scott Kirvan <Scott.Kirvan@intel.com>
Reviewed-on: https://review.whamcloud.com/29900
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>