CART-89 group: Tidy up some unclear rc handling code.
Rework some code that appeared to be checking the return of
crt_group_rank() but was actually checking the value of rc.
Remove calls to crt_group_rank() that weren't required
Remove the rc variable and access it via the callback handle directly.
Replace a if/trigraph asignment with a if/else if loop for readability.

This is intended to improve readability, behaviour shouldn't be altered.

Remove checks on the return values of crt_req_get() and crt_reply_get().

Change-Id: I1093a45ed46a01d566e62e6e27ade8a0ee887143
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29917
Reviewed-by: Yulu Jia <yulu.jia@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>