CORCI-159 Fix/Simplify update_devel_branch.sh
The sanity checks that update_devel_branch was doing to try to validate
the commit hashes in the build.config were not working, producing both
false positives and negatives.

The fix is to assume that the build.config is correct and then make sure
that the component branches are properly set.

Change-Id: I65bc28fe160b60f7f28eee8e8b055decf0ad7338
Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>
Reviewed-on: https://review.whamcloud.com/30017
Tested-by: Jenkins
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>