DAOS-340 vos: allow VOS to use DRAM
Current VOS can run in DRAM mode only when it's loaded by test
tool, this patch allows VOS to run in DRAM mode integrated with
the server stack. It can be enabled by env:

VOS_MEM_CLASS="DRAM"

In this mode, all VOS data and metadata are volatile because
they are stored in DRAM. The major intention of this mode is
allowing developer to evaluate latency of DAOS software stack.

Change-Id: I49232f966848075b66998c9d308825f8c9df1750
Signed-off-by: Liang Zhen <liang.zhen@intel.com>
Reviewed-on: https://review.whamcloud.com/30125
Reviewed-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>