DAOS-414 progress: Avoid waiting if timeout is zero
When crt_progress() is called with a zero timeout, crt_hg_progress()
forces a 1 ms timeout at the Mercury level. This patch moves the 1 ms
restriction from crt_hg_timeout() to crt_progress(), so that

  - when crt_progress() is called with a zero timeout, we don't wait;
  - when crt_progress() is called with a nonzero timeout, we wait at the
    Mercury level for 1 ms to keep current semantics (avoiding busy
    polling).

Note that this patch does not fix the case when crt_progress() is called
with a nonzero timeout and a NULL cond_cb. In this case, we enter the
"fast path" and wait at the Mercury level, giving up the chance to check
in-flight RPC timeouts.

Change-Id: Id554b370f86d42c33b3583eacfd4ff74cfb12393
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.whamcloud.com/30238
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Xuezhao Liu <xuezhao.liu@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>