CART-361 group: Have the resample RPC hold a reference on group.
If the LM code sends a resample for a remote rank then hold a reference
on the group until the RPC completes.  This avoids the case where
the group is detached before the RPC completes.

Whilst this solves the immediate issue I think a more complete fix
would be to have group_detach evict all RPCs to group ranks.

Change-Id: I4f848141b1b1a4f847822c268f93415089c35e22
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30424
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Xuezhao Liu <xuezhao.liu@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>