CART-355 tests: Transfer IV fetched data back to client
This patch adds support for transferring (via BULK_PUT) the results of
an IV fetch operation back to the command line client. The client then
prints that data to STDOUT as properly formatted JSON text so that it
can be easily parsed by a higher-level test wrapper utility.

All other output from the IV client should now go to STDERR to make
parsing the output easier.

Change-Id: Ifc6f70f2087a87d50a2bc86956a84cb0480c232e
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/30679
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>