DAOS-487 build: Use new go builder
Adds a symbolic link from site_scons to scons_local.  This is
a bit of a short term hack as we decide what to do with site
tools.   It may be that we add a package install step from
scons_local to site_scons to the build.

We may also move prereq_tools and build_info to be scons tools
to simplify the SConstruct a bit.

This uses GoProgram builder for the go file which should behave
more like Program (building intermediate object files, etc)

The update to scons_local also fixes an issue with
--update-prereq option
Also fixes an issue with scons_local/utils/setup_local.sh which
can make use of .build_vars.sh to setup environment

Change-Id: I865b5c3bf89714142d87b1244ce5800b45022526
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/30998
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>