DAOS-487 build: Add go builder
The base of this code comes from here:

https://bitbucket.org/zombiezen/scons-go-tools/overview

However, that code doesn't work with new go releases so
I've updated it.  Tested with daos build and it seems to
work.

To use this, one can create a symbolic link from
site_scons to scons_local (or perhaps move scons_local to
site_scons...but that is for another day).

Then just
env = Environment(TOOLS=['default', 'go'])

and

env.GoProgram(name, sources)

Change-Id: I21ebd9eb5a550cfa1218eab4effabe731b17f46d
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/31006
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>
Reviewed-by: John Malmberg <john.e.malmberg@intel.com>