IOF-343 fuse: Fix fuse build
Rename fusermount3 binary after installing.
As the build does not set user or suid bits on the suid binary it
cannot be used to launch a fuse filesystem.  Rename the new binary
so it doesn't take precedence over an installed one.

It would be possible to simply remove this binary instead but renameing
it allows users to run the chmod options directly if they want.

Change-Id: Iccffff7f8ea38ed96cf9208b676d7f2f3eab2fa0
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31072
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>