DAOS-331 build: Move the PMDK commit to latest master
Landed a new patch in the PMDK library to improve their build
to pick-up ARM64 architectures and build automatically.

Also PMDK has introduced a library NDTCL which seems to fail
builds if not found. Currently disabling this until PMDK fixes.
scons_local has been modified, so scons_local is updated to master
in this patch.

Added a couple lines in README to show that DAOS is supported
in ARM platforms also.

Change-Id: Iab9964d5c88725df32ff3acb832785a5e41162ae
Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-on: https://review.whamcloud.com/31077
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>