DAOS-382 test: Create test tool & add tests
This patch includes a new test and/or user tool daosctl that performs
DAOS management actions.

It also performs functions that only make sense for testing. It is
presently tied into the overall DAOS scons build system.

There is some limited help included in the tool which can be invoked
by daosctl --help.

Change-Id: I50efa996575ebb33a71c031f74ce84702a0f29f2
Signed-off-by: Scott Kirvan <scott.kirvan@intel.com>
Reviewed-on: https://review.whamcloud.com/31189
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Liang Zhen <liang.zhen@intel.com>