DAOS-515 vos: Add metadata attribute for failed objects
Enables rebuild to mark an object as failed so higher layers
can skip failed shards.  This currently places the
responsibility on the user to check the value.   This can
likely be improved.

Adds tests for 3 new APIs, one to set metadata flags, one
to clear metadata flags, and another to read metadata flags.

I had to disable a few vos tests that are failing.  Filed
DAOS-558

Change-Id: I656087a3f64d290d902817fba1f2c87028752e0a
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/31368
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>