DAOS-555 btree: tr_at can point to invalid offset
If there are two btree contexts accessing the same
btree, tr_at might be larger than nd->tn_keyn,
for example

1. Rebuild thread(A) is scanning the OI table by
at=11, and key_number=19 on tr_node Ta(leaf node).
2. Creation thread(B) is inserting new entry into
the table, which cause Ta to split into Ta_l(10) and
Ta_r(9).
3. Then A goes back and continue iterate from 11 on
Ta(because tr_at is in memory, B will not be able
to update it), but ta's key_number has become to 10,
so the iteration will not be ended.

Let's ending the iteration once tr_at >= key number.

Change-Id: Ieffc97c651014fd86f125a8816969588265dc26e
Signed-off-by: Wang Di <di.wang@intel.com>
Reviewed-on: https://review.whamcloud.com/31358
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Liang Zhen <liang.zhen@intel.com>