DAOS-458 common: Checksum library for DAOS
DAOS internal checksum library. Designed to work in both
ARM and X86 platforms. Uses ISA-L for x86 and uses mchecksum
for ARM processor. Added a simple test for the APIs alone.

Checksum usage is still in VOS, a separate patch will move this
to object layer and use the offload interface.

Change-Id: I0fd646a1f50ee0dc0a4d3d2804feb88dd470d149
Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-on: https://review.whamcloud.com/31179
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Liang Zhen <liang.zhen@intel.com>