DAOS-36: Add support for direct key comparison in btree
Store a pointer to the relevant leaf node in the space used
for hkey so we can do direct key comparison rather than hashed
key comparison.
Add a test case btree_direct.c that defines a tree class which
is sorted alphabetically.
Slight modification to how classes are initialized as we can't
assume BTR_FEAT_UINT_KEY anymore when hash callbacks are not
defined.

Change-Id: I05dd464e797b47dffdc0781a4b80921e5a2cf0e2
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/31682
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>