IOF-832 lm: Improve error handling in group_psr() function.
Create the new list as a local variable and only pass this back to
the caller on success, performing cleanup internally to this function
if there is any kind of failure.  Change the API so the call fails if
no PSRs are found.

This simplies the logic required by the caller, and fixes numerous
exit paths that would leave part-initialised rank lists.

Change-Id: I51bc69c02d74554997248487e4ffc76fe8b86807
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31725
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Yulu Jia <yulu.jia@intel.com>