DAOS-919 evtree: Simple insert, clip on fetch
Remove clipping on insertion in EVTree and instead
return clipped rectangles on fetch.   Interface is modified
to optionally return all rectangles or to get all
rectangles and then sort rectangles into two lists, one
for visible extents and another for rectangles that
are fully covered in the specified range.
Fix a compiler warning picked up by Jenkins.  Not sure why
a new warning made it through.

Change-Id: I63acab682c8eb8d67ae7728983a669233ff0e07c
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/31731
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeffrey Boyd <jeffrey.boyd@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>