DAOS-753 vos_tests: Fix 3 VOS I/O tests that had been failing
3 fixed tests are: io_iter_test_dkey_cond, io_obj_forward_recx_iter_test,
and io_obj_reverse_recx_iter_test. The first was failing because of a
null-pointer dereference: kbund.kb_epr was not set, but was dereferenced.
This is fixed in vos_obj.c.

The other two tests were fixed by removing a call to
io_update_and_fetch_dkey() in vts_io::io_obj_recx_range_iteration().
That call would update a dkey with epoch epr.epr_lo, but a later loop
would update dkeys with an epoch with a lesser value than epr.epr_lo.
Writing to an "older" epoch was the cause of the problem.

Change-Id: I4eb559ee1501a88b70c92d31657d73a4fa22f710
Signed-off-by: Jeff Boyd <jeffrey.boyd@intel.com>
Reviewed-on: https://review.whamcloud.com/31853
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>