DAOS-474 build: install DPDK libs along with SPDK libs
SPDK is dependent on few DPDK libs. There is a submodule of DPDK in
the SPDK repository, and it can be used by SPDK if no external DPDK
is specified when building SPDK (which is DAOS use case).

The problem is that SPDK 'make install' won't install the DPDK libs
built from the submodule, so we have to manually copy these DPDK libs
to the SPDK install directory.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I40bbafed66bb547ef5187c1347f8a1a72e89277e
Reviewed-on: https://review.whamcloud.com/32005
Tested-by: Jenkins
Reviewed-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>