DAOS-451 rdb: Rename rdb_tree.c to rdb_kvs.c
The KVS cache submodule is called rdb_tree, since each cache object maps
a KVS path to the matching dbtree handle. As rdb moves from dbtree to
vos, trees are replaced by VOS objects. This patch is the first of the
two steps to rename rdb_tree to rdb_kvs, a name independent of the
storage. (Two steps are required for Git to recognize the file rename.)

Change-Id: I02548bc6a8a6fea4d4d6558abb432917d23781a5
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.whamcloud.com/32232
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Liang Zhen <liang.zhen@intel.com>