DAOS-895 build: Add GOPATH/bin to the path and load the protobuf tool
In order for the protobuf tool to know where its components are it needs to be
loaded later in the SConstruct file. We add the GOPATH/bin to the path so we
have access to protoc-gen-go and then load the tool on demand.

Additionally update the scons_local reference to version that suports protobuf
compilation and GOPATH path variable and update the requirements to specify the
need to build the protobuf compiler.

Change-Id: I007c39e16e322825efdae7573645082195b23909
Signed-off-by: David Quigley <david.quigley@intel.com>
Reviewed-on: https://review.whamcloud.com/32490
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>