DAOS-839 test: test rebuild when capacity is insufficient
This test creates a pool with 6 targets, it writes data using mpio until
the pool has less than 1 targets worth of space remaining.  At that point
a target is failed/excluded.  This triggers a rebuild with the expectation
that the rebuild stops when it sees there is insufficient space to
complete.

Change-Id: I0f19a2e48573216061d8d5209218f909dd5c97f9
Signed-off-by: Scott Kirvan <scott.kirvan@intel.com>
Reviewed-on: https://review.whamcloud.com/32527
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Stephen Willson <stephen.d.willson@intel.com>
Reviewed-by: wangdi <di.wang@intel.com>