CART-511 perf: Remove version and magic from RPC header
None of these fields are used currently so simply remove them.
There is a flags field which seems to be a bit-field, the only bit being
the COLL_RPC bit which is a attribute of the RPC, so could be read from the
opc struct.
That leaves opcode which is needed for send, and rc which is needed
for reply, so it might be possible to combine these as well which would
require more work.

Change-Id: I1b400f52de4b6239e40fe8a1bfa7e5ba3646aee4
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32591
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>