DAOS-994: Remove protobuf/protoc requirement
Having the DAOS build generate the source from proto files turned out to be too
much of a burden on the developers and is really only required for people
manipulating the proto files. This patch removes the protobuf prereq and the
loading of the protoc tool. It does not revert the commit that added them
because the commit also bumped the scons_local version which is still needed.

Change-Id: I6af9ee42d98912a1f5313a290f9280b2a0883a36
Signed-off-by: David Quigley <david.quigley@intel.com>
Reviewed-on: https://review.whamcloud.com/32604
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>