CART-89 style: Do not pass bare variables to D_GOTO()
Pass in 0 rather than rc to D_GOTO() to avoid the impression
that this variable is somehow used by the macro.  The parameter
is executed, however if it is a bare variable this has no
effect so make it clear that nothing is happening in this call.

This commit updates gurt and the test code but does not yet
address the same issue in cart itself.

Change-Id: I64e329cf135275eee5f9db875e493fc7497d28a3
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32616
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>