CART-329 batch: Improve the protocol registration interface.
Add the base opcode to the protocol struct directly to avoid
passing it around seperately.
Move RPC registration from a enum in crt_rpc.h and a array in
crt_rpc.c to a single X-macro list in crt_rpc.h to clarify
RPC definitions and avoid consistency problems.
Now that opcodes are defined in one place remove the _internal
version of the protocol structs and registration functions,
just use the public registration function for the internal
protocol.
Remove the proto_type enum now internal and external are treated
the same.

- Test change to make rpc array/enum list synched up

Change-Id: Id88a6ae745c5c8b4842fc6a28856a7a701a5d32c
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32643
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Yulu Jia <yulu.jia@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>