DAOS-1069 Test: Fix an out of date daos_test flag comment.
daos_test has an out of date comment that lists 'd' must come
last. But then the default doesn't even follow that, as it now
has 'r' after it. Really, just all tests that destroy nodes need
to be last, which now includes rebuild, etc.

Update the comment to be more generic and accurate.

Change-Id: Ic066a8a8aeea0b7c7eff2037a99b7c46e8e5f39b
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32774
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Stephen Willson <stephen.d.willson@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Scott Kirvan <scott.kirvan@intel.com>