DAOS-1332 Display errno when fallocate() fails
When fallocate() fails in tgt_vos_create() it's useful to know why.

Display the string representation of the errno value in the error
message returned when fallocate() fails.

Also, get rc from daos_errno2der(errno) rather than using the the
return value from fallocate() which will always be -1 in this case.

Change-Id: I85bdccc1af391f96bf10985ce5a529ab39ae9204
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33140
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>