CORCI-338 Add tool to lookup dependencies.
utils/SConstruct_info: A script that generates a shell script
with the generic information needed for using a system that
can pre-build the components.

Usage:
   pushd daos
   scons -f scons_local/utils/SConstruct_info \
       --output-script=info.sh \
       --build-config=utils/build.config
   popd

The --build-config defaults to utils/build.config if not specified.

The output-script is sourced or read by another script for processing
the pre-built components.

Note that this query is only done for the architecture that the script
is run on, so when run on x86_64 will not show the ARM components.

Change-Id: Ia3d7c21afcd391c6417a12d396953f7a1ec0bd1a
Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32786
Tested-by: Jenkins
Reviewed-by: Phyllis Crandall <phyllis.crandall@intel.com>
Tested-by: Phyllis Crandall <phyllis.crandall@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>