CORCI-347 Use tag names to find pre-builts
Modify docker_host_postrun.sh and publish_coral_artifacts.sh to check to
see if the commit hash used for the build is referenced by a tag.

If it is referenced by a tag, save the tag and use that tag as a
symbolic link for the build.

Once we have run a new set of builds with this change, it should be
possible to use tags in the build.config instead of commit hashes when a
specific release is wanted.

Fix find in docker_host_build_components.sh and cart_post_build.sh to
only return the first matched file.

Change-Id: I2233459d5f183eda31f81c85568aeb857ef4fc06
Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32826
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>