DAOS-1074 test: Add rdb tests to CI
Currently the rdb tests are run manually with the command line commands:
server:
orterun -N 1 --enable-recovery --report-uri /tmp/urifile <debug_cmds>
daos_server -c 1 -m vos,rdb,rdbt

client:
orterun --ompi-server file:/tmp/urifile <debug_cmds> -np 1 rdbt init
--group=daos_server --uuid <uuid>
orterun --ompi-server file:/tmp/urifile <debug_cmds> -np 1 rdbt test --update
--group=daos_server
orterun --ompi-server file:/tmp/urifile <debug_cmds> -np 1 rdbt test
--group=daos_server
orterun --ompi-server file:/tmp/urifile <debug_cmds> -np 1 rdbt fini
--group=daos_server

Where debug_cmds = -x D_LOG_MASK=DEBUG,RPC=ERR,MEM=ERR -x DD_SUBSYS=all
-x DD_MASK=all

This patch is a python script that automates the process and is called from
utils/run_tests.sh.

A separate process is created (multiprocessing.Process) that starts
the DAOS server. If the server is started successfully, the output
from that command is saved to a file since it contains the PID of
the daos_io_server. At the end of the run, we kill the daos_io_server,
daos_server, and orterun processes using this PID.  If the server fails to
start, the parent detects this by testing the exit code before starting
the client operations. If this code is not None, the server has failed
to start. The parent returns an error code to the test framework.

In the original process the client commands are issued and if an error
is detected, the server is shut down with kill -9 <PID> and an error
code is returned. Otherwise, the tests run to completion and
signal SUCCESS and a 0 return code.

Regardless of the way in which the run ended, success or failure, the server is
shut down and the file that contains info for the child PID is deleted.

Change-Id: I08b704b5274ca1d5f616e192f7b837df3ae11e72
Signed-off-by: Phyllis Crandall <phyllis.crandall@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32899
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>