DAOS-1167 test: Fix issues with evtree unit test
The evtree test hasn't actually been working since
evt_insert_sgl was removed.   The evt_ctl.sh script
wasn't passing the return code back to the run_test.sh
script and, therefore, the test was failing silently.

Now the negative tests that are in it are now working
correctly and the full test can be reenabled.

This is a bit of a hack.   I faked the eio_addr_t but it
appears to work by simply telling evt_insert that it's SCM
even though, in reality, it isn't.  We can potentially
improve this but I wanted to get the basic test to run
again.

Also fixes linting issues with evt_ctl.sh since we
started checking shell scripts

Change-Id: I7f8ef90c24d1bba2fd33d5e3095d7d01df52241f
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32928
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>