DAOS-1173 test: Fix btree test to fail if pool creation fails
It took me a bit to figure out why my btree tests were failing
and it turned out it was simply a lack of space.   Since we weren't
checking the return value of pmemobj_create, it just silently
continued and seg faulted when we tried to use it.

Also fixes btree.sh script to get rid of shellcheck warnings

Change-Id: If966158f147e8edb83e7945ebe4c83f14b58a090
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32935
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>