CART-491 gurt: Avoid debug messages preparation if debug is disabled
There is a lot of time a code spend for arguments preparation even if
debug is disabled. For example, the function CP_UUID() is widely used
in debug macros and consume a lot of CPU time.

Move check for debug enabled inside the D_DEBUG macro from d_log() to
avoid arguments preparation if debug is disabled.

The performance of following command on my laptop:
daos_perf -T vos -P 1G -d 1 -a 200 -r 1000 -s 1k -C 0 -t -z

with original code is the following:
    duration : 8.212142   sec
    bandwith : 23.226     MB/sec
    rate     : 24354.18   IO/sec
    latency  : 41.061     us (nonsense if credits > 1)

with this patch become the following:
    duration : 5.645295   sec
    bandwith : 33.787     MB/sec
    rate     : 35427.73   IO/sec
    latency  : 28.226     us (nonsense if credits > 1)

Change-Id: I6eabb3102e76f620f2d7f7de468257effd878d21
Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32943
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>