DAOSM-39 gurt: user can register alternative assert
This allows a user to register an alternative assert function. This is
useful for unit testing. For example, cmocka has the ability to test if
a function throws an assertion, but to do this, the user must replace
calls to assert() with mock_assert() from the cmocka library. Using a
function pointer here means the module that calls d_register_alt_assert()
is the only one that needs to import the cmocka library.

Change-Id: I6b2bcf1f12743a64e3969b194ac0a46df947413f
Signed-off-by: Jeff Boyd <jeffrey.boyd@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32957
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>