DAOS-1208 test: Fix rdb test
Right after the rdb tests were enabled to run in the Jenkins build CI a
change was made to rdb that caused the tests to fail (a change in screen
output when starting the daos server). This patch is a restructuring of
the rdb_test_runner so that screen output is no longer used to determine
the PID for the daos server.

These tests are run by issuing an orterun command to start a daos
server and more orterun commands to run the client tests. At the end
of the test sequence it is necessary to kill the daos_server process,
otherwise it would continue into eternity. This new incarnation uses
subprocess.Popen to start the daos server rather than os.system. Popen
allows control to return to the original process after the orterun command
is given and polling can determine if the child process is s till alive.

The method for finding the PID of the daos_server so it can be killed
is essentially this: ps -o pid,comm --no-headers --ppid <PID for parent
process> gives a ps list of all children of that process. For each child,
recursively look at their children using ps --ppid until one named daos_*
is found.  Capture that PID and kill it.

As a side issue, code has been added that checks that, even if the tests
all passed, if the server crashed before the end of the run, an rc of
1 is returned to the test framework.

Signed-off-by: Phyllis Crandall <phyllis.crandall@intel.com>
Change-Id: I14e57efce61da4cba92710e2c5c1a34f67955755
Reviewed-on: https://review.hpdd.intel.com/32999
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins