DAOS-1228 evtree: Add function for getting maximum visible offset
In order to find the end of the evtree range, one has to iterate
through the range.  This is expensive enough to do inside of evtree
without introducing multiple round trip messages.  This takes a
simple approach of repeating the search until either we find
an unpunched maximum or we find nothing.

Change-Id: Ia4d7e463869a7cdedb962e66574c17d58c708d02
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33026
Reviewed-by: Jeffrey Boyd <jeffrey.boyd@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>