DAOS-1606 evtree: revert to original full overwrite
In commit d5cc39d93ea06b722ca494085280fbd3c5973735 , the evtree full
overwrite was changed to free the passed in region represented by the
new evt_ptr directly instead of replacing the old evt_ptr by the new
one.

This optimization is actually buggy, because the passed in region
represented by new evt_ptr (could be SCM or NVMe) is just being
reserved but not published yet, so we should call cancel instead of
free on it, furthermore, the reserved list on caller side needs be
adjusted, otherwise, the new region will be mistakenly published on
update end.

It's not quite easy to adjust the reserved list inside of evtree, so I
think we'd just revert back to original copy approach for this moment.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I22ce8b9ba19ada3f65d4b493d36dc5375e21b254
Reviewed-on: https://review.hpdd.intel.com/33408
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>